requirements.txt
openai>=1.54.0
python-telegram-bot==20.7
python-dotenv>=1.0.1
python-docx>=1.1.0
numpy>=1.26.4
openpyxl>=3.1.5
